requirements.txt
--extra-index-url=http://dist.pinaxproject.com/dev/
--extra-index-url=http://dist.pinaxproject.com/alpha/

Django<2
django-extensions
pinax-theme-bootstrap==2.0.4
pinax-theme-bootstrap-account==1.0b2
django-appconf==1.0.2
webtest
django-webtest
pinax-notifications
toposort
django-cors-headers
rdflib
rdflib-jsonld
djangorestframework
jsonrpclib
django-contrib-comments
html5lib>0.99999999
bleach==2.0.0
django-simple-captcha
django-mptt
django-fobi==0.12.2
django-modeltranslation<0.14
pytz
graphql-core==1.1
graphene==1.4
graphene-django==1.3
PyJWT
six
selenium
mock
passlib
bcrypt
